Add typerighter helper, and move transforms into helper folder
