+++ src/elements/helpers/typerighter.ts
@@ -1,0 +1,9 @@
+import type { NodeSpec } from "prosemirror-model";
+
+export const useTyperighterNodeSpec: Partial<NodeSpec> = {
+  attrs: {
+    useTyperighter: {
+      default: true,
+    },
+  },
+};
