+++ src/elements/image/imageElementDataTransformer.ts
@@ -1,5 +1,5 @@
 import type { FieldNameToValueMap } from "../../plugin/helpers/fieldView";
-import type { TransformIn, TransformOut } from "../transformer/types/Transform";
+import type { TransformIn, TransformOut } from "../helpers/types/Transform";
 import type { Asset, createImageFields, MainImageData } from "./ImageElement";
 import { undefinedDropdownValue } from "./ImageElement";
